Renamed some tests
