+++ tests/Feature/NewsItemShowTest.php
@@ -6,10 +6,9 @@ use App\Models\Category;
 use App\Models\News;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class ShowCategoryItemTest extends TestCase
+class NewsItemShowTest extends TestCase
 {
     use RefreshDatabase;
     /**
@@ -17,23 +16,25 @@ class ShowCategoryItemTest extends TestCase
      *
      * @return void
      */
-    public function test_show_category_is_success()
+    public function test_show_news_is_success()
     {
         $user = User::factory()->create();
         $this->actingAs($user);
 
         $category = Category::factory()->create();
 
-        $response = $this->get("/categories/{$category->id}");
+        $news = News::factory()->create(['category_id' => $category->id]);
+
+        $response = $this->get("/news/{$news->id}");
         $response->assertStatus(200);
     }
 
-    public function test_show_category_with_fake_id_is_error()
+    public function test_show_news_with_fake_id_is_error()
     {
         $user = User::factory()->create();
         $this->actingAs($user);
 
-        $response = $this->get("/categories/1");
+        $response = $this->get("/news/1");
         $response->assertStatus(404);
     }
 }
